+++ docs/whats-new.rst
@@ -12,8 +12,9 @@ Breaking changes
 
 Enhancements
 ~~~~~~~~~~~~
-- Add ``cook_rgidf()`` in ``utils._funcs`` to make it easy of 
-  using a non-RGI glacier inventory in OGGM (:pull:`1251`).
+
+- Added ``cook_rgidf()`` function in ``oggm.utils`` to simplify the use
+  of a non-RGI glacier inventory in OGGM (:pull:`1251`).
   By `Li Fei <https://github.com/Keeptg>`_
 
 Bug fixes